[ADD]: 다른 사람 정보 보기, 관리자만 보기 기능 구현

+++ backend/service/user/src/main/java/backend/service/user/service/UserServiceImpl.java
@@ -12,8 +12,8 @@ import backend.service.user.entity.UserEntity;
 import backend.service.user.feignClient.BoardClient;
 import backend.service.user.feignClient.CommentClient;
 import backend.common.jwt.JwtUtil;
-import backend.service.user.kafka.KafkaProducer;
 import backend.service.user.repository.UserRepository;
+import backend.service.user.util.SecurityUtil;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
@@ -31,7 +31,6 @@ public class UserServiceImpl implements UserService {
     private final UserRepository userRepository;
     private final BCryptPasswordEncoder encoder;
     private final JwtUtil jwtUtil;
-    private final KafkaProducer kafkaProducer;
 
     private final BoardClient boardClient;
     private final CommentClient commentClient;
@@ -39,24 +38,42 @@ public class UserServiceImpl implements UserService {
     @Override
     public CreateResponse create(CreateRequest dto) {
 
-        UserEntity entity = userRepository.save(UserEntity.create(snowflake.nextId(), dto.getUsername(), encoder.encode(dto.getPassword()), dto.getEmail()));
+        UserEntity entity = userRepository.save(UserEntity.create(snowflake.nextId(), dto.getUsername(), encoder.encode(dto.getPassword()), dto.getEmail(), dto.getAdminPassword()));
 
         return CreateResponse.from(entity);
     }
 
     @Override
-    public UpdateResponse update(UpdateRequest dto, Long userId) {
+    public UpdateResponse update(UpdateRequest dto) {
+
+        Long userId = SecurityUtil.getCurrentUserId();
+
         UserEntity entity = userRepository.findByUserId(userId);
-        log.info(userId);
+
+        if (entity == null) {
+            throw new RuntimeException("유저 없음");
+        }
+
         entity.update(dto.getUsername(), dto.getPassword(), dto.getEmail());
+
         userRepository.save(entity);
+
         return UpdateResponse.from(entity);
     }
 
     @Override
-    public DeletedResponse delete(DeleteRequest dto, Long userId) {
+    public DeletedResponse delete(DeleteRequest dto) {
+
+        Long userId = SecurityUtil.getCurrentUserId();
+
         UserEntity entity = userRepository.findByUserId(userId);
+
+        if (entity == null) {
+            throw new RuntimeException("유저 없음");
+        }
+
         entity.delete();
+
         return DeletedResponse.from(entity);
     }
 
@@ -81,7 +98,6 @@ public class UserServiceImpl implements UserService {
         return LoginResponse.from(entity, accessToken);
     }
 
-
     @Override
     public List<CreateResponse> getAllUsers() {
         List<UserEntity> entities = userRepository.findAll();
@@ -89,7 +105,9 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public GetUserResponse getUser(Long userId) {
+    public GetMyInfoResponse getMyInfo() {
+
+        Long userId = SecurityUtil.getCurrentUserId();
 
         UserEntity entity = userRepository.findByUserId(userId);
 
@@ -100,7 +118,36 @@ public class UserServiceImpl implements UserService {
         List<BoardDto> responseBoards = boardClient.getBoards(userId);
         List<CommentDto> responseComments = commentClient.getComments(userId);
 
+        return GetMyInfoResponse.from(entity, responseBoards, responseComments);
+    }
+
+    @Override
+    public GetUserResponse getUser(Long userId) {
+
+        UserEntity entity = userRepository.findByUserId(userId);
+
+        if (entity == null) {
+            throw new RuntimeException("존재하지 않는 사용자 입니다.");
+        }
+        List<BoardDto> responseBoards = boardClient.getBoards(userId);
+        List<CommentDto> responseComments = commentClient.getComments(userId);
+
         return GetUserResponse.from(entity, responseBoards, responseComments);
+    }
+
+    @Override
+    public GetUserResponse getUserByAdmin(Long userId) {
+
+        UserEntity entity = userRepository.findByUserId(userId);
+
+        if (entity == null) {
+            throw new RuntimeException("존재하지 않는 사용자 입니다.");
+        }
+
+        List<BoardDto> boards = boardClient.getBoards(userId);
+        List<CommentDto> comments = commentClient.getComments(userId);
+
+        return GetUserResponse.from(entity, boards, comments);
     }
 
 }
